[LH-56] Separate Text and Image class type #5
