+++ data/src/main/java/com/lighthouse/android/data/api/interceptor/ContentInterceptor.kt
@@ -27,6 +27,6 @@ object ContentInterceptor : Interceptor {
     }
 
     private fun getContentsList(json: JsonObject): JsonArray {
-        return json.getAsJsonArray("contents")
+        return json.getAsJsonObject("data").getAsJsonArray("content")
     }
 }
