+++ domain/src/main/java/com/lighthouse/domain/response/RichText.kt
@@ -1,11 +1,20 @@
 package com.lighthouse.domain.response
 
+
 data class RichText(
+    val textRichType: RichTextType?,
+    val imageRichType: RichImageType?
+)
+
+data class RichTextType(
     val text: String?,
     val style: List<String>?,
     val textColor: String?,
     val background: String?,
     val size: Float?,
+)
+
+data class RichImageType(
     val url: String?,
     val width: Float?,
     val height: Float?
